Move wmap to sysproc
