+++ xv6-public/usys.S
@@ -31,3 +31,4 @@ SYSCALL(sleep)
 SYSCALL(uptime)
 SYSCALL(getpgdirinfo)
 SYSCALL(getwmapinfo)
+SYSCALL(wmap)
